+++ api/layer/container.go
@@ -13,7 +13,6 @@ import (
 	"github.com/nspcc-dev/neofs-sdk-go/container"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	"github.com/nspcc-dev/neofs-sdk-go/eacl"
-	"github.com/nspcc-dev/neofs-sdk-go/session"
 	session2 "github.com/nspcc-dev/neofs-sdk-go/session/v2"
 	"go.uber.org/zap"
 )
@@ -143,11 +142,10 @@ func (n *layer) createContainer(ctx context.Context, p *CreateBucketParams) (*da
 		Creator:              bktInfo.Owner,
 		Policy:               p.Policy,
 		Name:                 p.Name,
-		SessionToken:         p.SessionContainerCreation,
 		SessionTokenV2:       p.SessionTokenV2,
 		CreationTime:         bktInfo.Created,
 		AdditionalAttributes: attributes,
-	}, *p.EACL, p.SessionEACL, p.SessionTokenV2)
+	}, *p.EACL, p.SessionTokenV2)
 	if err != nil {
 		return nil, fmt.Errorf("create container: %w", err)
 	}
@@ -161,10 +159,10 @@ func (n *layer) createContainer(ctx context.Context, p *CreateBucketParams) (*da
 	return bktInfo, nil
 }
 
-func (n *layer) setContainerEACLTable(ctx context.Context, idCnr cid.ID, table *eacl.Table, sessionToken *session.Container, sessionTokenV2 *session2.Token) error {
+func (n *layer) setContainerEACLTable(ctx context.Context, idCnr cid.ID, table *eacl.Table, sessionTokenV2 *session2.Token) error {
 	table.SetCID(idCnr)
 
-	err := n.neoFS.SetContainerEACL(ctx, *table, sessionToken, sessionTokenV2)
+	err := n.neoFS.SetContainerEACL(ctx, *table, sessionTokenV2)
 	if err == nil {
 		n.cache.PutBucketACL(idCnr, table)
 	}