+++ api/layer/notifications.go
@@ -9,7 +9,6 @@ import (
 	"github.com/nspcc-dev/neofs-s3-gw/api"
 	"github.com/nspcc-dev/neofs-s3-gw/api/data"
 	"github.com/nspcc-dev/neofs-s3-gw/api/s3headers"
-	"github.com/nspcc-dev/neofs-sdk-go/session"
 	session2 "github.com/nspcc-dev/neofs-sdk-go/session/v2"
 )
 
@@ -20,17 +19,13 @@ type PutBucketNotificationConfigurationParams struct {
 }
 
 func (n *layer) PutBucketNotificationConfiguration(ctx context.Context, p *PutBucketNotificationConfigurationParams) error {
-	var (
-		sessionToken   *session.Container
-		sessionTokenV2 *session2.Token
-	)
+	var sessionTokenV2 *session2.Token
 	boxData, err := GetBoxData(ctx)
 	if err == nil {
-		sessionToken = boxData.Gate.SessionTokenForSetAttribute()
 		sessionTokenV2 = boxData.Gate.SessionTokenV2
 	}
 
-	if err = n.storeAttribute(ctx, p.BktInfo.CID, attributeNotifications, p.Configuration, sessionToken, sessionTokenV2); err != nil {
+	if err = n.storeAttribute(ctx, p.BktInfo.CID, attributeNotifications, p.Configuration, sessionTokenV2); err != nil {
 		return fmt.Errorf("store bucket notification settings: %w", err)
 	}
 
@@ -77,7 +72,7 @@ func (n *layer) GetBucketNotificationConfiguration(ctx context.Context, bktInfo 
 	boxData, err := GetBoxData(ctx)
 	if err == nil {
 		// Migrate notification settings.
-		if err = n.storeAttribute(ctx, bktInfo.CID, attributeNotifications, conf, boxData.Gate.SessionTokenForSetAttribute(), boxData.Gate.SessionTokenV2); err != nil {
+		if err = n.storeAttribute(ctx, bktInfo.CID, attributeNotifications, conf, boxData.Gate.SessionTokenV2); err != nil {
 			return nil, fmt.Errorf("migrate bucket notification settings: %w", err)
 		}
 		if err = n.deleteBucketMetaObjects(ctx, bktInfo, s3headers.TypeBucketNotifConfig); err != nil {