+++ api/layer/versioning_test.go
@@ -9,13 +9,22 @@ import (
 	"github.com/nspcc-dev/neofs-s3-gw/api"
 	"github.com/nspcc-dev/neofs-s3-gw/api/data"
 	"github.com/nspcc-dev/neofs-s3-gw/creds/accessbox"
-	bearertest "github.com/nspcc-dev/neofs-sdk-go/bearer/test"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
+	session2 "github.com/nspcc-dev/neofs-sdk-go/session/v2"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 )
+
+func newTestSessionTokenV2(t *testing.T, signer user.Signer) *session2.Token {
+	var tok session2.Token
+	tok.SetVersion(session2.TokenCurrentVersion)
+	tok.SetIssuer(signer.UserID())
+	require.NoError(t, tok.SetSubjects([]session2.Target{session2.NewTargetUser(signer.UserID())}))
+	require.NoError(t, tok.Sign(signer))
+	return &tok
+}
 
 func (tc *testContext) putObject(content []byte) *data.ObjectInfo {
 	extObjInfo, err := tc.layer.PutObject(tc.ctx, &PutObjectParams{
@@ -142,13 +151,9 @@ func prepareContext(t *testing.T, cachesConfig ...*CachesConfig) *testContext {
 
 	signer := user.NewAutoIDSignerRFC6979(key.PrivateKey)
 
-	bearerToken := bearertest.Token()
-	require.NoError(t, bearerToken.Sign(signer))
-
 	ctx := context.WithValue(context.Background(), api.BoxData, &accessbox.Box{
 		Gate: &accessbox.GateData{
-			BearerToken: &bearerToken,
-			GateKey:     key.PublicKey(),
+			SessionTokenV2: newTestSessionTokenV2(t, signer),
 		},
 	})
 	tp := NewTestNeoFS(signer)